+++ Sources/HIGThemesContract/HIGTheme.swift
@@ -72,6 +72,7 @@ public protocol HIGTheme: Sendable {
     var mediaRow: any HIGMediaRowTokens { get }
     var hero: any HIGHeroTokens { get }
     var listGroup: any HIGListGroupTokens { get }
+    var barChart: any HIGBarChartTokens { get }
     var photoPicker: any HIGPhotoPickerTokens { get }
     var photoEditor: any HIGPhotoEditorTokens { get }
     var longTextEditor: any HIGLongTextEditorTokens { get }